Fixed bugs with the flood fill tool.  Prevents overwriting pixels with same colour but different index and overflow bug when trying to flood fill area with same colour.

+++ src/com/plus/mevanspn/BBCSpriteEd/ui/panels/ImagePanel.java
@@ -181,7 +181,7 @@ final public class ImagePanel extends JPanel implements MouseListener, MouseMoti
                     repaint(this.getVisibleRect());
                     parent.UpdateTimeline();
                 } else if (isActiveDrawingTool("floodfill")) {
-                    activeImage.FloodFill(p, parent.GetActiveColourIndex(), 0, false);
+                    activeImage.FloodFill(p, parent.GetActiveColourIndex(), (byte) 0, false);
                     repaint(this.getVisibleRect());
                     parent.UpdateTimeline();
                 } else if (isActiveDrawingTool("paintbrush")) {
